// src/lib/oidc-provider/types.ts
export interface CodeVerificationValue {
  /**
   * The client ID
   */
  clientId: string;
  /**
   * The redirect URI for the client
   */
  redirectURI: string;
  /**
   * The scopes that the client requested
   */
  scope: string[];
  /**
   * The user ID
   */
  userId: string;
  /**
   * The time that the user authenticated
   */
  authTime: number;
  /**
   * Whether the user needs to consent to the scopes
   * before the code can be exchanged for an access token.
   *
   * If this is true, then the code is treated as a consent
   * request. Once the user consents, the code will be updated
   * with the actual code.
   */
  requireConsent: boolean;
  /**
   * The state parameter from the request
   *
   * If the prompt is set to `consent`, then the state
   * parameter is saved here. This is to prevent the client
   * from using the code before the user consents.
   */
  state: string | null;
  /**
   * Code challenge
   */
  codeChallenge?: string;
  /**
   * Code Challenge Method
   */
  codeChallengeMethod?: 'sha256' | 'plain';
  /**
   * Nonce
   */
  nonce?: string;
}